+++ lib/unit_measurements/unit_groups/all.rb
@@ -17,8 +17,10 @@ require_relative "weight"
 require_relative "area"
 require_relative "volume"
 require_relative "density"
+require_relative "force"
 
 ## Other units
-require_relative "plane_angle"
 require_relative "sound_level"
+require_relative "plane_angle"
+require_relative "solid_angle"
 require_relative "quantity"